+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 *   New feature: improved API to accessing `resolve` objects. They are now available
     via the `routing` HOC or the `Router` FaC, alongside the `location`props
-*   Huge improvments and additions to docs
+*   Huge improvements and additions to docs
 
 ## v1.0.0-beta.1